Change styling for the body container component and moved it into the atoms of the assets folder
