+++ src/assets/styles/atoms/BodyContainer.js
@@ -1,9 +1,8 @@
 import styled from "styled-components";
-import { media } from "../index";
+import { media } from "../variables/media";
 
 const BodyContainer = styled.div`
-  width: 100%;
-  height: 100%;
+  width: 100%; height: 100%;
   display: flex;
   align-items: center;
   justify-content: ${({ justify }) => justify ? justify : "start"};
@@ -12,13 +11,6 @@ const BodyContainer = styled.div`
   @media ${media.tablet} {
     flex-direction: column;
     padding: 60px 20px;
-    & > img {
-      display: none;
-    }
-  }
-
-  @media ${media.mobile} {
-    flex-direction: column;
     & > img {
       display: none;
     }
