+++ src/assets/styles/views/PageNotFoundStyling.js
@@ -1,4 +1,4 @@
-import BodyContainer from "../../../components/atoms/BodyContainer";
+import BodyContainer from "../atoms/BodyContainer";
 import { H2 } from "../../../components/atoms/Heading";
 import styled from "styled-components";
