+++ src/components/organisms/FormLayout.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import WideBody from "../atoms/WideBody";
-import BodyContainer from "../atoms/BodyContainer";
+import BodyContainer from "../../assets/styles/atoms/BodyContainer";
 import HeroImage from "../atoms/HeroImage";
 import Form from "./Form";
